+++ src/Interfaces/ManyRelationInterface.php
@@ -3,6 +3,8 @@
 namespace Swis\JsonApi\Client\Interfaces;
 
 use Swis\JsonApi\Client\Collection;
+use Swis\JsonApi\Client\Links;
+use Swis\JsonApi\Client\Meta;
 
 interface ManyRelationInterface
 {
@@ -39,4 +41,24 @@ interface ManyRelationInterface
      * @return bool
      */
     public function shouldOmitIncluded(): bool;
+
+    /**
+     * @param \Swis\JsonApi\Client\Links|null $links
+     */
+    public function setLinks(Links $links = null);
+
+    /**
+     * @return \Swis\JsonApi\Client\Links|null
+     */
+    public function getLinks();
+
+    /**
+     * @param \Swis\JsonApi\Client\Meta|null $meta
+     */
+    public function setMeta(Meta $meta = null);
+
+    /**
+     * @return \Swis\JsonApi\Client\Meta|null
+     */
+    public function getMeta();
 }