+++ src/JsonApi/ErrorsParser.php
@@ -4,15 +4,30 @@ namespace Swis\JsonApi\Client\JsonApi;
 
 use Art4\JsonApiClient\Error as JsonApiError;
 use Art4\JsonApiClient\ErrorCollection as JsonApiErrorCollection;
+use Art4\JsonApiClient\ErrorLink as JsonApiErrorLink;
 use Art4\JsonApiClient\ErrorSource as JsonApiErrorSource;
 use Art4\JsonApiClient\Meta as JsonApiMeta;
 use Swis\JsonApi\Client\Errors\Error;
 use Swis\JsonApi\Client\Errors\ErrorCollection;
-use Swis\JsonApi\Client\Errors\ErrorMeta;
 use Swis\JsonApi\Client\Errors\ErrorSource;
+use Swis\JsonApi\Client\Links;
+use Swis\JsonApi\Client\Meta;
 
 class ErrorsParser
 {
+    /**
+     * @var \Swis\JsonApi\Client\JsonApi\LinksParser
+     */
+    private $linksParser;
+
+    /**
+     * @param \Swis\JsonApi\Client\JsonApi\LinksParser $linksParser
+     */
+    public function __construct(LinksParser $linksParser)
+    {
+        $this->linksParser = $linksParser;
+    }
+
     /**
      * @param \Art4\JsonApiClient\ErrorCollection $errorCollection
      *
@@ -21,14 +36,8 @@ class ErrorsParser
     public function parse(JsonApiErrorCollection $errorCollection)
     {
         $errors = new ErrorCollection();
-        $errorCollectionArray = $errorCollection->asArray(false);
 
-        // Empty errors
-        if (empty($errorCollectionArray)) {
-            throw new \InvalidArgumentException('Error collection does not contain any errors.');
-        }
-
-        foreach ($errorCollectionArray as $error) {
+        foreach ($errorCollection->asArray(false) as $error) {
             $errors->push($this->buildError($error));
         }
 
@@ -44,13 +53,24 @@ class ErrorsParser
     {
         return new Error(
             $error->has('id') ? $error->get('id') : null,
+            $error->has('links') ? $this->buildLinks($error->get('links')) : null,
             $error->has('status') ? $error->get('status') : null,
             $error->has('code') ? $error->get('code') : null,
             $error->has('title') ? $error->get('title') : null,
             $error->has('detail') ? $error->get('detail') : null,
             $error->has('source') ? $this->buildErrorSource($error->get('source')) : null,
-            $error->has('meta') ? $this->buildErrorMeta($error->get('meta')) : null
+            $error->has('meta') ? $this->buildMeta($error->get('meta')) : null
         );
+    }
+
+    /**
+     * @param \Art4\JsonApiClient\ErrorLink $errorLink
+     *
+     * @return \Swis\JsonApi\Client\Links
+     */
+    private function buildLinks(JsonApiErrorLink $errorLink): Links
+    {
+        return $this->linksParser->parse($errorLink->asArray(false));
     }
 
     /**
@@ -69,10 +89,10 @@ class ErrorsParser
     /**
      * @param \Art4\JsonApiClient\Meta $meta
      *
-     * @return \Swis\JsonApi\Client\Errors\ErrorMeta
+     * @return \Swis\JsonApi\Client\Meta
      */
-    private function buildErrorMeta(JsonApiMeta $meta): ErrorMeta
+    private function buildMeta(JsonApiMeta $meta): Meta
     {
-        return new ErrorMeta($meta->asArray(false));
+        return new Meta($meta->asArray(true));
     }
 }