+++ src/Providers/ServiceProvider.php
@@ -17,6 +17,7 @@ use Swis\JsonApi\Client\Interfaces\ParserInterface;
 use Swis\JsonApi\Client\Interfaces\TypeMapperInterface;
 use Swis\JsonApi\Client\JsonApi\ErrorsParser;
 use Swis\JsonApi\Client\JsonApi\Hydrator;
+use Swis\JsonApi\Client\JsonApi\LinksParser;
 use Swis\JsonApi\Client\JsonApi\Parser;
 use Swis\JsonApi\Client\TypeMapper;
 
@@ -58,8 +59,9 @@ class ServiceProvider extends BaseServiceProvider
             function (Application $app) {
                 return new Parser(
                     new JsonApiClientManger(),
-                    new Hydrator($app->make(TypeMapperInterface::class)),
-                    new ErrorsParser()
+                    new Hydrator($app->make(TypeMapperInterface::class), new LinksParser()),
+                    new ErrorsParser(new LinksParser()),
+                    new LinksParser()
                 );
             }
         );