+++ tests/JsonApi/HydratorTest.php
@@ -7,6 +7,10 @@ use Art4\JsonApiClient\Utils\Manager;
 use Swis\JsonApi\Client\Collection;
 use Swis\JsonApi\Client\Item;
 use Swis\JsonApi\Client\JsonApi\Hydrator;
+use Swis\JsonApi\Client\JsonApi\LinksParser;
+use Swis\JsonApi\Client\Link;
+use Swis\JsonApi\Client\Links;
+use Swis\JsonApi\Client\Meta;
 use Swis\JsonApi\Client\Relations\HasOneRelation;
 use Swis\JsonApi\Client\Relations\MorphToManyRelation;
 use Swis\JsonApi\Client\Relations\MorphToRelation;
@@ -43,7 +47,7 @@ class HydratorTest extends AbstractTest
      */
     protected function getHydrator(): Hydrator
     {
-        return new Hydrator($this->getTypeMapperMock());
+        return new Hydrator($this->getTypeMapperMock(), new LinksParser());
     }
 
     /**
@@ -86,11 +90,20 @@ class HydratorTest extends AbstractTest
                             'type' => 'child',
                             'id'   => '2',
                         ],
+                        'links' => [
+                            'self' => 'http://example.com/'.$type.'/'.$id.'/relationships/child',
+                        ],
+                        'meta' => [
+                            'foo' => 'bar',
+                        ],
                     ],
                     'morph'     => [
                         'data' => [
                             'type' => 'child',
                             'id'   => '3',
+                        ],
+                        'meta' => [
+                            'foo' => 'bar',
                         ],
                     ],
                     'morphmany' => [
@@ -108,7 +121,16 @@ class HydratorTest extends AbstractTest
                                 'id'   => '6',
                             ],
                         ],
+                        'links' => [
+                            'self' => 'http://example.com/'.$type.'/'.$id.'/relationships/morphmany',
+                        ],
                     ],
+                ],
+                'links' => [
+                    'self' => 'http://example.com/master/1',
+                ],
+                'meta' => [
+                    'foo' => 'bar',
                 ],
             ],
             'included' => [
@@ -209,7 +231,7 @@ class HydratorTest extends AbstractTest
         $typeMapper = new TypeMapper();
         $typeMapper->setMapping('child', ChildItem::class);
         $typeMapper->setMapping('master', MasterItem::class);
-        $hydrator = new Hydrator($typeMapper);
+        $hydrator = new Hydrator($typeMapper, new LinksParser());
 
         $childJsonApiItem = $this->getJsonApiChildItemMock(2, 'master', 1);
         $childItem = $hydrator->hydrateItem($childJsonApiItem);
@@ -226,11 +248,15 @@ class HydratorTest extends AbstractTest
 
         static::assertInstanceOf(MasterItem::class, $masterItem);
         static::assertInstanceOf(HasOneRelation::class, $masterItem->getRelationship('child'));
+        static::assertInstanceOf(Links::class, $masterItem->getRelationship('child')->getLinks());
+        static::assertInstanceOf(Meta::class, $masterItem->getRelationship('child')->getMeta());
 
         static::assertSame($childItem, $masterItem->getRelationship('child')->getIncluded());
         static::assertEquals('child', $masterItem->getRelationship('child')->getIncluded()->getType());
         static::assertEquals(2, $masterItem->getRelationship('child')->getIncluded()->getId());
         static::assertSame($masterItem, $masterItem->getRelationship('child')->getIncluded()->getRelationship('parent')->getIncluded());
+        static::assertSame('http://example.com/master/1/relationships/child', $masterItem->getRelationship('child')->getLinks()->self->getHref());
+        static::assertSame('bar', $masterItem->getRelationship('child')->getMeta()->foo);
     }
 
     /**
@@ -242,7 +268,7 @@ class HydratorTest extends AbstractTest
         /** @var \Swis\JsonApi\Client\Interfaces\TypeMapperInterface $typeMapper */
         $typeMapper = new TypeMapper();
         $typeMapper->setMapping('master', MasterItem::class);
-        $hydrator = new Hydrator($typeMapper);
+        $hydrator = new Hydrator($typeMapper, new LinksParser());
 
         $data = [
             'data' => [
@@ -345,7 +371,7 @@ class HydratorTest extends AbstractTest
         $typeMapper = new TypeMapper();
         $typeMapper->setMapping('child', ChildItem::class);
         $typeMapper->setMapping('master', MasterItem::class);
-        $hydrator = new Hydrator($typeMapper);
+        $hydrator = new Hydrator($typeMapper, new LinksParser());
 
         $childJsonApiItem = $this->getJsonApiChildItemMock(3);
         $childItem = $hydrator->hydrateItem($childJsonApiItem);
@@ -376,7 +402,7 @@ class HydratorTest extends AbstractTest
         $typeMapper = new TypeMapper();
         $typeMapper->setMapping('child', ChildItem::class);
         $typeMapper->setMapping('master', MasterItem::class);
-        $hydrator = new Hydrator($typeMapper);
+        $hydrator = new Hydrator($typeMapper, new LinksParser());
 
         $childJsonApiItem = $this->getJsonApiChildItemMock(4);
         $childItem = $hydrator->hydrateItem($childJsonApiItem);
@@ -406,7 +432,7 @@ class HydratorTest extends AbstractTest
         /** @var \Swis\JsonApi\Client\Interfaces\TypeMapperInterface $typeMapper */
         $typeMapper = new TypeMapper();
         $typeMapper->setMapping('item-without-relationships', WithoutRelationshipsItem::class);
-        $hydrator = new Hydrator($typeMapper);
+        $hydrator = new Hydrator($typeMapper, new LinksParser());
 
         $childJsonApiItem = $this->getJsonApiChildItemMock(3);
         $childItem = $hydrator->hydrateItem($childJsonApiItem);
@@ -436,7 +462,7 @@ class HydratorTest extends AbstractTest
         /** @var \Swis\JsonApi\Client\Interfaces\TypeMapperInterface $typeMapper */
         $typeMapper = new TypeMapper();
         $typeMapper->setMapping('item-without-relationships', WithoutRelationshipsItem::class);
-        $hydrator = new Hydrator($typeMapper);
+        $hydrator = new Hydrator($typeMapper, new LinksParser());
 
         $childJsonApiItem = $this->getJsonApiChildItemMock(4);
         $childItem = $hydrator->hydrateItem($childJsonApiItem);
@@ -455,5 +481,35 @@ class HydratorTest extends AbstractTest
         static::assertEquals(4, $masterItem->getRelationship('morphmany')->getIncluded()[0]->getId());
 
         static::assertEquals(4, $masterItem->morphmany[0]->getId());
+    }
+
+    /**
+     * @test
+     */
+    public function it_hydrates_links()
+    {
+        $hydrator = $this->getHydrator();
+
+        $jsonApiItem = $this->getJsonApiItemMock('master', 1);
+        $item = $hydrator->hydrateItem($jsonApiItem);
+
+        static::assertInstanceOf(Links::class, $item->getLinks());
+
+        static::assertEquals(new Links(['self' => new Link('http://example.com/master/1')]), $item->getLinks());
+    }
+
+    /**
+     * @test
+     */
+    public function it_hydrates_meta()
+    {
+        $hydrator = $this->getHydrator();
+
+        $jsonApiItem = $this->getJsonApiItemMock('master', 1);
+        $item = $hydrator->hydrateItem($jsonApiItem);
+
+        static::assertInstanceOf(Meta::class, $item->getMeta());
+
+        static::assertEquals(new Meta(['foo' => 'bar']), $item->getMeta());
     }
 }